fix: 🐛 move back to channel route
